initial merge with portal, no user interface yet

+++ grails-app/domain/au/org/emii/portal/User.groovy
@@ -57,7 +57,7 @@ class User {
 
     static def current = {
         def currentSubject = SecurityUtils.getSubject()
-        def principal = currentSubject?.getPrincipal()
+        def principal = currentSubject?.getPrincipal() as Integer
         if (principal) {
             def userInstance = User.get(principal)
